// src/dto/control-menu-dto.ts
export interface IGeneralControlMenu {
    productionPlan: boolean;
    preparation: boolean;
    productionProblem: boolean;
    dataDefect: boolean;
    method: boolean;
    material: boolean;
    equipment: boolean;
    manpower: boolean;
    layout: boolean;
}

export interface IProductionPlanMenu {
    schedule: boolean;
    equipment: boolean;
    manpower: boolean;
    finish: boolean;
}